校区删除优化 返回值优化

+++ src/main/java/com/child/programming/education/manage/controller/SchoolController.java
@@ -4,10 +4,13 @@ import com.child.programming.base.dto.LoginedUserInfoDto;
 import com.child.programming.base.dto.ResultDto;
 import com.child.programming.base.dto.SchoolInfoDto;
 import com.child.programming.base.model.TbSchoolDo;
+import com.child.programming.base.service.IClassroomService;
 import com.child.programming.base.service.ISchoolService;
+import com.child.programming.base.util.EmptyUtils;
 import com.child.programming.base.util.HttpSessionUtil;
 import com.child.programming.base.util.ResponseUtil;
 import com.child.programming.education.manage.dto.SchoolInfoSelectDto;
+import com.child.programming.education.manage.dto.ValidateClassroomInfoDto;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -22,19 +25,20 @@ import java.util.List;
  * @Author：yangfan
  **/
 
-@Controller
+@RestController
 @RequestMapping("/school")
 @Log4j2
 public class SchoolController {
     @Autowired
     private ISchoolService iSchoolService;
+    @Autowired
+    private IClassroomService iClassroomService;
 
     /**
      * 查询校区
      * @return
      */
     @RequestMapping(value = "getList", method = RequestMethod.GET)
-    @ResponseBody
     public List<SchoolInfoDto> getList(@RequestParam(value = "name", required = false) String name){
         return iSchoolService.getList(name);
     }
@@ -47,7 +51,6 @@ public class SchoolController {
      * @return
      */
     @RequestMapping(value = "save", method = RequestMethod.POST)
-    @ResponseBody
     public ResultDto save(HttpSession session, @RequestBody TbSchoolDo schoolDo){
         LoginedUserInfoDto userInfoPojo = HttpSessionUtil.getLoginedUserInfo(session);
         if (null != userInfoPojo && null != schoolDo){
@@ -56,7 +59,7 @@ public class SchoolController {
                 return ResultDto.success();
         }
 
-        return ResultDto.error(ResponseUtil.ERROR_MSG);
+        return ResultDto.fail();
     }
 
     /**
@@ -66,7 +69,6 @@ public class SchoolController {
      * @return
      */
     @RequestMapping(value = "delete", method = RequestMethod.GET)
-    @ResponseBody
     public ResultDto delete(@RequestParam(value = "idsStr", required = true)String idsStr,
                             HttpSession session) {
         log.info(idsStr + "删除");
@@ -74,11 +76,17 @@ public class SchoolController {
         LoginedUserInfoDto userInfoPojo = HttpSessionUtil.getLoginedUserInfo(session);
         if (null != userInfoPojo && !StringUtils.isEmpty(idsStr)) {
             String[] idArray = idsStr.split(",");
-            boolean result = iSchoolService.delete(idArray, userInfoPojo.getId());
-            if (result)
-                return ResultDto.success();
+            List<ValidateClassroomInfoDto> validateClassroomInfoDtoList = iClassroomService.validateSchoolId(idArray);
+            //没有教室占用，可以删除
+            if (EmptyUtils.listIsEmpty(validateClassroomInfoDtoList)){
+                boolean result = iSchoolService.delete(idArray, userInfoPojo.getId());
+                if (result)
+                    return ResultDto.success();
+            }
+            //有教室占用，不能删除
+            return ResultDto.fail(ResponseUtil.FAIL_MSG, validateClassroomInfoDtoList);
         }
-        return ResultDto.error(ResponseUtil.ERROR_MSG);
+        return ResultDto.fail();
     }
 
     /**
@@ -86,8 +94,8 @@ public class SchoolController {
      * @return
      */
     @RequestMapping("getSchoolInfoSelect")
-    @ResponseBody
     public List<SchoolInfoSelectDto> getSchoolInfoSelect(){
         return iSchoolService.getSchoolInfoSelectList();
     }
+
 }
